Move a significant amount of compute logic to shade

+++ shade_ansible/spec.py
@@ -26,7 +26,7 @@ def openstack_argument_spec(**kwargs):
         project_domain_name=dict(default=None),
         state=dict(default='present', choices=['absent', 'present']),
         wait=dict(default=True, type='bool'),
-        timeout=dict(default=180, type='float'),
+        timeout=dict(default=180, type='int'),
         endpoint_type=dict(
             default='publicURL', choices=['publicURL', 'internalURL']
         )
